335. Fixing Invoice Links

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const Product = require('../models/product')
-const { validationResult } = require('express-validator');
+const {validationResult} = require('express-validator');
+const {deleteFile} = require("../util/file");
 
 exports.getAdminProducts = (req, res, next) => {
 
@@ -31,10 +32,28 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
     const errors = validationResult(req);
+    console.log(image)
+    if (!image) {
+        return res.status(422).render('admin/edit-product', {
+            pageTitle: 'Add Product',
+            path: '/admin/edit-product',
+            editing: false,
+            hasError: true,
+            product: {
+                title: title,
+                price: price,
+                description: description
+            },
+            errorMessage: 'Attached file is not a image !',
+            validationErrors: errors.array()
+        });
+    }
+
+    const imageUrl = image.path;
 
     if (!errors.isEmpty()) {
         console.log(errors.array());
@@ -107,7 +126,8 @@ exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const updatedTitle = req.body.title;
     const updatedPrice = req.body.price;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
+    // const updatedImageUrl = req.body.image;
     const updatedDesc = req.body.description;
 
     const errors = validationResult(req);
@@ -120,7 +140,6 @@ exports.postEditProduct = (req, res, next) => {
             hasError: true,
             product: {
                 title: updatedTitle,
-                imageUrl: updatedImageUrl,
                 price: updatedPrice,
                 description: updatedDesc,
                 _id: prodId
@@ -138,13 +157,17 @@ exports.postEditProduct = (req, res, next) => {
             product.title = updatedTitle;
             product.price = updatedPrice;
             product.description = updatedDesc;
-            product.imageUrl = updatedImageUrl;
+            if (image) {
+                deleteFile(product.imageUrl)
+                product.imageUrl = image.path;
+            }
             return product.save().then(result => {
                 console.log('UPDATED PRODUCT!');
                 res.redirect('/admin/products');
             });
         })
         .catch(err => {
+            console.log(err)
             const error = new Error(err)
             error.statusCode = 500;
             return next(error)
@@ -155,7 +178,12 @@ exports.postEditProduct = (req, res, next) => {
 exports.postDeleteProduct = (req, res, next) => {
     const {productId} = req.body;
 
-    Product.deleteOne({_id: productId, userId: req.user._id})
+    Product.findById(productId)
+        .then(product => {
+            if (!product) return next(new Error('Product Not Found!'))
+            deleteFile(product.imageUrl)
+            return Product.deleteOne({_id: productId, userId: req.user._id});
+        })
         .then(result => {
             res.redirect('/admin/products');
         })
